refactor(lib2): extract colored frame creation to class
